broke up db.rs into init.rs, ops.rs and window.rs
