+++ src-tauri/src/db/mod.rs
@@ -1,0 +1,3 @@
+pub mod ops;
+pub mod init;
+pub mod window;
