+++ src-tauri/src/main.rs
@@ -3,7 +3,7 @@
 
 // Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
 
-use habtrack::db::{init::DbInitializer, ops, DB_SINGLETON};
+use habtrack::db::{init::DbInitializer, ops, window};
 
 // Start work on database.
 // TODO: Once done, merge with main, and pull changes into FEATURE-add-delete-task.
@@ -12,7 +12,7 @@ fn main() {
     tauri::Builder::default()
         .setup(|app| {
             let db_init_obj = DbInitializer::new(app.handle()).init();
-            DB_SINGLETON
+            ops::DB_SINGLETON
                 .lock()
                 .unwrap()
                 .set_conn(db_init_obj.get_db_path().as_str())
@@ -25,8 +25,8 @@ fn main() {
             ops::crud_commands::add_item,
             ops::crud_commands::delete_item,
             ops::crud_commands::update_item,
-            ops::app_commands::open_tomorrow_window,
-            ops::app_commands::close_tomorrow_window,
+            window::open_tomorrow_window,
+            window::close_tomorrow_window,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
